Move useUniqueId to a more general folder
