+++ public/app/features/search/page/components/FolderSection.tsx
@@ -8,7 +8,7 @@ import { Card, Checkbox, CollapsableSection, Icon, Spinner, useStyles2 } from '@
 import { config } from 'app/core/config';
 import { t } from 'app/core/internationalization';
 import { getSectionStorageKey } from 'app/features/search/utils';
-import { useUniqueId } from 'app/plugins/datasource/influxdb/components/useUniqueId';
+import { useUniqueId } from 'app/hooks/useUniqueId';
 
 import { SearchItem } from '../..';
 import { GENERAL_FOLDER_UID } from '../../constants';
